+++ pkg/goboot/goboot.go
@@ -12,6 +12,7 @@ import (
 	"path/filepath"
 
 	"github.com/it-timo/goboot/pkg/baseci"
+	"github.com/it-timo/goboot/pkg/basegovernance"
 	"github.com/it-timo/goboot/pkg/basedocker"
 	"github.com/it-timo/goboot/pkg/baselint"
 	"github.com/it-timo/goboot/pkg/baselocal"
@@ -198,6 +199,11 @@ func (gb *GoBoot) registerMainServices() error {
 			err := gb.ServiceMgr.register(baserelease.NewBaseRelease(gb.cfg.TargetPath))
 			if err != nil {
 				return fmt.Errorf("failed to register %s service: %w", goboottypes.ServiceNameBaseRelease, err)
+			}
+		case goboottypes.ServiceNameBaseGovernance:
+			err := gb.ServiceMgr.register(basegovernance.NewBaseGovernance(gb.cfg.TargetPath))
+			if err != nil {
+				return fmt.Errorf("failed to register %s service: %w", goboottypes.ServiceNameBaseGovernance, err)
 			}
 		// Future services can be added here.
 		default: