+++ pkg/goboottypes/names.go
@@ -21,4 +21,6 @@ const (
 	ServiceNameBaseDocker = "base_docker"
 	// ServiceNameBaseRelease identifies the release automation generator.
 	ServiceNameBaseRelease = "base_release"
+	// ServiceNameBaseGovernance identifies the governance generator.
+	ServiceNameBaseGovernance = "base_governance"
 )